Move MaybeBlock, ConstSize, GetExtra types to primitives.
