+++ pallets/committee/src/benchmarking.rs
@@ -21,11 +21,9 @@ use frame_support::{
     StorageValue,
 };
 use frame_system::RawOrigin as SystemOrigin;
-use polymesh_common_utilities::{
-    benchs::{user, User},
-    MaybeBlock,
-};
+use polymesh_common_utilities::benchs::{user, User};
 use polymesh_primitives::committee::COMMITTEE_MEMBERS_MAX;
+use polymesh_primitives::MaybeBlock;
 use sp_std::prelude::*;
 
 const PROPOSAL_PADDING_WORDS: usize = 1_000;
